+++ fixtures/ts/eslint/eslintrc.js
@@ -1,28 +1,35 @@
+/* eslint-env node */
 module.exports = {
   root: true,
   env: {
     browser: true,
     es6: true,
-    node: true,
   },
-  plugins: ['prettier'],
-  extends: ['eslint:recommended', 'plugin:prettier/recommended', 'prettier'],
+  parser: '@typescript-eslint/parser',
+  plugins: ['@typescript-eslint', 'prettier'],
+  extends: [
+    'eslint:recommended',
+    'plugin:@typescript-eslint/eslint-recommended',
+    'plugin:@typescript-eslint/recommended',
+    'plugin:prettier/recommended',
+    'prettier',
+  ],
   rules: {
-    'no-console': process.env.NODE_ENV === 'production' ? 'error' : 'off',
-    'no-debugger': process.env.NODE_ENV === 'production' ? 'error' : 'off',
     quotes: [2, 'single', 'avoid-escape'],
     semi: [2, 'never'],
     'no-extra-boolean-cast': 'off',
     'no-unused-vars': [
       1,
       {
+        ignoreRestSiblings: true,
         argsIgnorePattern: 'res|next|^err',
       },
     ],
+    '@typescript-eslint/no-non-null-assertion': 'off',
     'prettier/prettier': [
       'error',
       {
-        trailingComma: 'es5',
+        trailingComma: 'all',
         singleQuote: true,
         tabWidth: 2,
         semi: false,